#494 - mu: map type framework
