+++ src/mu/core/mod.rs
@@ -10,6 +10,7 @@ pub mod exception;
 pub mod frame;
 pub mod functions;
 pub mod indirect;
+pub mod map;
 pub mod mu;
 pub mod namespace;
 pub mod reader;
