+++ src/mu/types/stream.rs
@@ -10,20 +10,17 @@
 use crate::{
     core::{
         exception::{self, Condition, Exception},
-        frame::Frame,
         indirect::IndirectTag,
         mu::{Core as _, Mu},
-        reader::{Core as _, Reader},
         types::{Tag, TagType, Type},
     },
     system::{stream::Core as _, sys::System},
     types::{
         char::Char,
         fixnum::Fixnum,
-        streambuilder::StreamBuilder,
         symbol::{Core as _, Symbol},
         vecimage::{TypedVec, VecType},
-        vector::{Core as _, Vector},
+        vector::Core as _,
     },
 };
 
@@ -80,7 +77,7 @@ impl Stream {
                 }
                 _ => panic!(),
             },
-            _ => panic!(),
+            _ => panic!("stream type botch {:?}", Tag::type_of(tag)),
         }
     }
 
@@ -154,7 +151,7 @@ impl Core for Stream {
     fn close(mu: &Mu, stream: Tag) {
         let mut image = Self::to_image(mu, stream);
 
-        System::close(&mu.system, Fixnum::as_i64(mu, image.stream_id) as usize).unwrap();
+        System::close(&mu.system, Fixnum::as_i64(image.stream_id) as usize).unwrap();
 
         image.stream_id = Symbol::keyword("t");
         Self::update(mu, &image, stream);
@@ -167,7 +164,7 @@ impl Core for Stream {
 
         let image = Self::to_image(mu, stream);
 
-        Ok(System::get_string(&mu.system, Fixnum::as_i64(mu, image.stream_id) as usize).unwrap())
+        Ok(System::get_string(&mu.system, Fixnum::as_i64(image.stream_id) as usize).unwrap())
     }
 
     fn write(mu: &Mu, tag: Tag, _: bool, stream: Tag) -> exception::Result<()> {
@@ -177,7 +174,7 @@ impl Core for Stream {
                 match Tag::type_of(image.stream_id) {
                     Type::Keyword => mu.write_string("#<stream: closed>".to_string(), stream),
                     Type::Fixnum => mu.write_string(
-                        format!("#<stream: id: {}>", Fixnum::as_i64(mu, image.stream_id)),
+                        format!("#<stream: id: {}>", Fixnum::as_i64(image.stream_id)),
                         stream,
                     ),
                     Type::Null | Type::Cons | Type::Vector => {
@@ -210,7 +207,7 @@ impl Core for Stream {
 
         match Tag::type_of(image.stream_id) {
             Type::Fixnum => {
-                let stream_id = Fixnum::as_i64(mu, image.stream_id) as usize;
+                let stream_id = Fixnum::as_i64(image.stream_id) as usize;
                 let unch = image.unch;
 
                 if unch.null_() {
@@ -253,7 +250,7 @@ impl Core for Stream {
 
         match Tag::type_of(image.stream_id) {
             Type::Fixnum => {
-                let stream_id = Fixnum::as_i64(mu, image.stream_id) as usize;
+                let stream_id = Fixnum::as_i64(image.stream_id) as usize;
                 let unch = image.unch;
 
                 if unch.null_() {
@@ -317,7 +314,7 @@ impl Core for Stream {
 
         match Tag::type_of(image.stream_id) {
             Type::Fixnum => {
-                let stream_id = Fixnum::as_i64(mu, image.stream_id) as usize;
+                let stream_id = Fixnum::as_i64(image.stream_id) as usize;
                 System::write_byte(&mu.system, stream_id, ch as u8)
             }
             _ => panic!(),
@@ -337,302 +334,13 @@ impl Core for Stream {
 
         match Tag::type_of(image.stream_id) {
             Type::Fixnum => {
-                let stream_id = Fixnum::as_i64(mu, image.stream_id) as usize;
+                let stream_id = Fixnum::as_i64(image.stream_id) as usize;
                 System::write_byte(&mu.system, stream_id, byte)
             }
             _ => panic!(
                 "internal: {:?} stream state inconsistency",
                 Tag::type_of(image.stream_id)
             ),
-        }
-    }
-}
-
-pub trait MuFunction {
-    fn mu_close(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_eof(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_flush(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_get_string(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_open(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_openp(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_read(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_read_byte(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_read_char(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_unread_char(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_write(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_write_byte(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-    fn mu_write_char(_: &Mu, _: &mut Frame) -> exception::Result<()>;
-}
-
-impl MuFunction for Stream {
-    fn mu_close(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-
-        fp.value = match Tag::type_of(stream) {
-            Type::Stream => {
-                if Self::is_open(mu, stream) {
-                    Self::close(mu, stream);
-                    Symbol::keyword("t")
-                } else {
-                    Tag::nil()
-                }
-            }
-            _ => return Err(Exception::new(Condition::Type, "close", stream)),
-        };
-
-        Ok(())
-    }
-
-    fn mu_openp(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-
-        fp.value = match Tag::type_of(stream) {
-            Type::Stream => {
-                if Self::is_open(mu, stream) {
-                    stream
-                } else {
-                    Tag::nil()
-                }
-            }
-            _ => return Err(Exception::new(Condition::Type, "openp", stream)),
-        };
-
-        Ok(())
-    }
-
-    fn mu_open(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let st_type = fp.argv[0];
-        let st_dir = fp.argv[1];
-        let st_arg = fp.argv[2];
-
-        let arg = match Tag::type_of(st_arg) {
-            Type::Vector => Vector::as_string(mu, st_arg),
-            _ => return Err(Exception::new(Condition::Type, "open", st_arg)),
-        };
-
-        let input = match Tag::type_of(st_dir) {
-            Type::Keyword if st_dir.eq_(Symbol::keyword("input")) => true,
-            Type::Keyword if st_dir.eq_(Symbol::keyword("output")) => false,
-            _ => return Err(Exception::new(Condition::Type, "open", st_dir)),
-        };
-
-        match Tag::type_of(st_type) {
-            Type::Keyword if st_type.eq_(Symbol::keyword("file")) => {
-                let stream = if input {
-                    StreamBuilder::new().file(arg).input().build(mu)
-                } else {
-                    StreamBuilder::new().file(arg).output().build(mu)
-                };
-
-                fp.value = match stream {
-                    Err(e) => return Err(e),
-                    Ok(stream) => stream.evict(mu),
-                };
-
-                Ok(())
-            }
-            Type::Keyword if st_type.eq_(Symbol::keyword("string")) => {
-                let stream = if input {
-                    StreamBuilder::new().string(arg).input().build(mu)
-                } else {
-                    StreamBuilder::new().string(arg).output().build(mu)
-                };
-
-                fp.value = match stream {
-                    Err(e) => return Err(e),
-                    Ok(stream) => stream.evict(mu),
-                };
-
-                Ok(())
-            }
-            _ => Err(Exception::new(Condition::Type, "open", st_type)),
-        }
-    }
-
-    fn mu_flush(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-
-        let image = Self::to_image(mu, stream);
-
-        fp.value = Tag::nil();
-
-        if !Self::is_open(mu, stream) {
-            return Ok(());
-        }
-
-        if image.direction.eq_(Symbol::keyword("input")) {
-            return Err(Exception::new(Condition::Stream, "flush", stream));
-        }
-
-        match Tag::type_of(image.stream_id) {
-            Type::Fixnum => {
-                let stream_id = Fixnum::as_i64(mu, image.stream_id) as usize;
-                System::flush(&mu.system, stream_id).unwrap()
-            }
-            _ => return Err(Exception::new(Condition::Type, "flush", stream)),
-        }
-
-        Ok(())
-    }
-
-    fn mu_read(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-        let eofp = fp.argv[1];
-        let eof_value = fp.argv[2];
-
-        match Tag::type_of(stream) {
-            Type::Stream => match Reader::read(mu, stream, !eofp.null_(), eof_value, false) {
-                Ok(tag) => {
-                    fp.value = tag;
-                    Ok(())
-                }
-                Err(e) => Err(e),
-            },
-            _ => Err(Exception::new(Condition::Type, "read", stream)),
-        }
-    }
-
-    fn mu_write(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let value = fp.argv[0];
-        let escape = fp.argv[1];
-        let stream = fp.argv[2];
-
-        match Tag::type_of(stream) {
-            Type::Stream => match mu.write(value, !escape.null_(), stream) {
-                Ok(_) => {
-                    fp.value = value;
-                    Ok(())
-                }
-                Err(e) => Err(e),
-            },
-            _ => Err(Exception::new(Condition::Type, "write", stream)),
-        }
-    }
-
-    fn mu_eof(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-
-        match Tag::type_of(stream) {
-            Type::Stream => {
-                fp.value = if Self::is_eof(mu, stream) {
-                    Symbol::keyword("t")
-                } else {
-                    Tag::nil()
-                };
-                Ok(())
-            }
-            _ => Err(Exception::new(Condition::Type, "eof", stream)),
-        }
-    }
-
-    fn mu_get_string(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-
-        match Tag::type_of(stream) {
-            Type::Stream => match Self::get_string(mu, stream) {
-                Ok(string) => {
-                    fp.value = Vector::from_string(&string).evict(mu);
-                    Ok(())
-                }
-                Err(e) => Err(e),
-            },
-            _ => Err(Exception::new(Condition::Type, "get-str", stream)),
-        }
-    }
-
-    fn mu_read_char(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-        let eoferrp = fp.argv[1];
-        let eof_value = fp.argv[2];
-
-        fp.value = match Tag::type_of(stream) {
-            Type::Stream => match Self::read_char(mu, stream) {
-                Ok(Some(ch)) => Char::as_tag(ch),
-                Ok(None) if eoferrp.null_() => eof_value,
-                Ok(None) => return Err(Exception::new(Condition::Eof, "rd-char", stream)),
-                Err(e) => return Err(e),
-            },
-            _ => return Err(Exception::new(Condition::Type, "rd-char", stream)),
-        };
-
-        Ok(())
-    }
-
-    fn mu_read_byte(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let stream = fp.argv[0];
-        let erreofp = fp.argv[1];
-        let eof_value = fp.argv[2];
-
-        fp.value = match Tag::type_of(stream) {
-            Type::Stream => match Self::read_byte(mu, stream) {
-                Ok(Some(byte)) => Fixnum::as_tag(byte as i64),
-                Ok(None) if erreofp.null_() => eof_value,
-                Ok(None) => return Err(Exception::new(Condition::Eof, "rd-byte", stream)),
-                Err(e) => return Err(e),
-            },
-            _ => return Err(Exception::new(Condition::Type, "rd-byte", stream)),
-        };
-
-        Ok(())
-    }
-
-    fn mu_unread_char(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let ch = fp.argv[0];
-        let stream = fp.argv[1];
-
-        match Tag::type_of(stream) {
-            Type::Stream => match Self::unread_char(mu, stream, Char::as_char(mu, ch)) {
-                Ok(Some(_)) => {
-                    panic!()
-                }
-                Ok(None) => {
-                    fp.value = ch;
-                    Ok(())
-                }
-                Err(e) => Err(e),
-            },
-            _ => Err(Exception::new(Condition::Type, "un-char", stream)),
-        }
-    }
-
-    fn mu_write_char(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let ch = fp.argv[0];
-        let stream = fp.argv[1];
-
-        match Tag::type_of(ch) {
-            Type::Char => match Tag::type_of(stream) {
-                Type::Stream => match Self::write_char(mu, stream, Char::as_char(mu, ch)) {
-                    Ok(_) => {
-                        fp.value = ch;
-                        Ok(())
-                    }
-                    Err(e) => Err(e),
-                },
-                _ => Err(Exception::new(Condition::Type, "wr-char", stream)),
-            },
-            _ => Err(Exception::new(Condition::Type, "wr-char", stream)),
-        }
-    }
-
-    fn mu_write_byte(mu: &Mu, fp: &mut Frame) -> exception::Result<()> {
-        let byte = fp.argv[0];
-        let stream = fp.argv[1];
-
-        match Tag::type_of(byte) {
-            Type::Fixnum if Fixnum::as_i64(mu, byte) < 256 => match Tag::type_of(stream) {
-                Type::Stream => {
-                    match Self::write_byte(mu, stream, Fixnum::as_i64(mu, byte) as u8) {
-                        Ok(_) => {
-                            fp.value = byte;
-                            Ok(())
-                        }
-                        Err(e) => Err(e),
-                    }
-                }
-                _ => Err(Exception::new(Condition::Type, "wr-byte", stream)),
-            },
-            Type::Fixnum => Err(Exception::new(Condition::Range, "wr-byte", byte)),
-            _ => Err(Exception::new(Condition::Type, "wr-byte", byte)),
         }
     }
 }
